+++ tests/e2e/src/common/mod.rs
@@ -153,6 +153,7 @@ impl TestEnvironment {
             memory_store: Some(Arc::new(self.memory_store.clone())),
             core_memory_store: Some(Arc::new(self.core_memory_store.clone())),
             rag_store: None,
+            embedding_service: None,
             langsmith: None,
         }
     }